added PowerUp3 Script
changed Lives Script
created new PowerUp3 Object

+++ Tank Tactics/Assets/Scripts/Lives.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Lives : MonoBehaviour
 {
-    public int lives;
+    public float lives;
 
     private void OnTriggerEnter(Collider other)
     {
